+++ test/test_helper.rb
@@ -3,14 +3,12 @@
 require 'simplecov'
 SimpleCov.start
 
-$LOAD_PATH.unshift File.expand_path('../lib', __dir__)
-require 'adequate_serialization'
-
 require 'rails/all'
 require 'sqlite3'
 require 'minitest/autorun'
 
-AdequateSerialization.hook_into_rails!
+$LOAD_PATH.unshift File.expand_path('../lib', __dir__)
+require 'adequate_serialization'
 
 ActiveRecord::Base.establish_connection(
   adapter: 'sqlite3',